Program.cs add I IController_Bestellung und ControllerBestellung() Wo fahlende elemente gipt kommentiert.
IControllerBestellung() erweitert
ControllerBestellung() erweitert
