--- HHBK_Chemicals_ERP_CS/ControllerBestellung.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace HHBK_Chemicals_ERP_CS
-{
-    class ControllerBestellung : IController_Bestellung
-    {
-        int IController_Bestellung.Andern(Bestellung bestellung, Bestellung newBestellung)
-        {
-            throw new NotImplementedException();
-        }
-
-        int IController_Bestellung.Anlegen(Bestellung bestellung)
-        {
-            throw new NotImplementedException();
-        }
-
-        int IController_Bestellung.Loeschen(Bestellung bestellung)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
