+++ HHBK_Chemicals_ERP_CS/IControllerBestellung.cs
@@ -8,6 +8,8 @@ namespace HHBK_Chemicals_ERP_CS
 {
     interface IController_Bestellung
     {
+        Model Model { set; }
+        View_Bestellung View_Bestellung { set; }
         int Anlegen(Bestellung bestellung);
         int Andern(Bestellung bestellung, Bestellung newBestellung);
         int Loeschen(Bestellung bestellung);
